Struct subdomain_enumerator::results::Results

pub struct Results {
    pub domains: Vec<String>,
    pub store: Vec<Arc<Mutex<HashSet<String>>>>,
}

A struct holding results of a subdomain enumeration. domain_list is the list of original domains that were supplied, eg "Facebook.com". store holds the subdomains found for each supplied domain, eg "messenger.facebook.com".

Fields

'domains' - A Vector of Strings holding the domains that were enumerated

'store' - A Vec<Arc<Mutex<HashSet>>> where each element is a list of subdomains corresponding to a domain in domains.

Fields
domains: Vec<String> store: Vec<Arc<Mutex<HashSet<String>>>>
Methods

impl Results

pub fn new(num_domains: usize, domains_list: Vec<String>) -> Self

Initiates a new Results instance.

pub fn print_subdomains(&self)

Print the associated subdomains found for each domain.